refactor: move llm into single packages
