+++ packages/llamaindex/package.json
@@ -30,12 +30,15 @@
     "@google/generative-ai": "0.12.0",
     "@grpc/grpc-js": "^1.11.1",
     "@huggingface/inference": "^2.8.0",
+    "@llamaindex/anthropic": "workspace:*",
     "@llamaindex/cloud": "workspace:*",
     "@llamaindex/core": "workspace:*",
+    "@llamaindex/deepinfra": "workspace:*",
     "@llamaindex/env": "workspace:*",
     "@llamaindex/groq": "workspace:*",
     "@llamaindex/ollama": "workspace:*",
     "@llamaindex/openai": "workspace:*",
+    "@llamaindex/portkey-ai": "workspace:*",
     "@mistralai/mistralai": "^1.0.4",
     "@mixedbread-ai/sdk": "^2.2.11",
     "@pinecone-database/pinecone": "^3.0.2",
@@ -62,7 +65,6 @@
     "openai": "^4.60.0",
     "papaparse": "^5.4.1",
     "pathe": "^1.1.2",
-    "portkey-ai": "0.1.16",
     "rake-modified": "^1.0.8",
     "string-strip-html": "^13.4.8",
     "tiktoken": "^1.0.15",
