+++ packages/llm/anthropic/src/index.ts
@@ -1,0 +1,14 @@
+export {
+  AnthropicAgent,
+  AnthropicAgentWorker,
+  type AnthropicAgentParams,
+} from "./agent";
+export {
+  ALL_AVAILABLE_ANTHROPIC_LEGACY_MODELS,
+  ALL_AVAILABLE_ANTHROPIC_MODELS,
+  ALL_AVAILABLE_V3_5_MODELS,
+  ALL_AVAILABLE_V3_MODELS,
+  Anthropic,
+  AnthropicSession,
+  type AnthropicAdditionalChatOptions,
+} from "./llm";
