+++ packages/llm/deepinfra/src/index.ts
@@ -1,0 +1,6 @@
+export {
+  DeepInfraEmbedding,
+  type DeepInfraEmbeddingResponse,
+  type InferenceStatus,
+} from "./embedding";
+export { DeepInfra } from "./llm";
